通知快完了

+++ src/main/java/life/majiang/community/dto/QuestionDTO.java
@@ -4,15 +4,14 @@ import life.majiang.community.model.User;
 import lombok.Data;
 
 @Data
-public class QuestionDTO
-{
-    private Integer id;
+public class QuestionDTO {
+    private Long id;
     private String title;
     private String description;
     private String tag;
     private Long gmtCreate;
     private Long gmtModified;
-    private Integer creator;
+    private Long creator;
     private Integer viewCount;
     private Integer commentCount;
     private Integer likeCount;
